+++ Engine.Brain/Entity/NP.cs
@@ -22,10 +22,10 @@ namespace Engine.Brain.Entity
         /// </summary>
         /// <param name="inputs"></param>
         /// <returns></returns>
-        public static float[] Normalize(byte[] inputs)
+        public static double[] Normalize(byte[] inputs)
         {
             int count = inputs.Length;
-            float[] normal = new float[inputs.Length];
+            double[] normal = new double[inputs.Length];
             for (int i = 0; i < count; i++)
                 normal[i] = inputs[i] / 255f;
             return normal;
@@ -35,10 +35,10 @@ namespace Engine.Brain.Entity
         /// </summary>
         /// <param name="inputs"></param>
         /// <returns></returns>
-        public static float[] Normalize(float[] inputs, float max)
+        public static double[] Normalize(double[] inputs, double max)
         {
             int count = inputs.Length;
-            float[] normal = new float[inputs.Length];
+            double[] normal = new double[inputs.Length];
             for (int i = 0; i < count; i++)
                 normal[i] = inputs[i] / max;
             return normal;
@@ -49,9 +49,9 @@ namespace Engine.Brain.Entity
         /// <param name="hotIndex"></param>
         /// <param name="hotLength"></param>
         /// <returns></returns>
-        public static float[] ToOneHot(int hotIndex, int hotLength)
+        public static double[] ToOneHot(int hotIndex, int hotLength)
         {
-            float[] oneHot = new float[hotLength];
+            double[] oneHot = new double[hotLength];
             for (int i = 0; i < hotLength; i++)
                 oneHot[i] = i == hotIndex ? 1 : 0;
             return oneHot;
@@ -90,16 +90,16 @@ namespace Engine.Brain.Entity
         /// </summary>
         /// <param name="batchSzie">样本数量</param>
         /// <returns></returns>
-        public static List<float> CreateLabels(int batchSzie = 15, int oneHot = 10)
+        public static List<double> CreateLabels(int batchSzie = 15, int oneHot = 10)
         {
-            var inputs = new List<float[]>();
+            var inputs = new List<double[]>();
             //构建多样本的输出label
             for (int i = 0; i < batchSzie; i++)
             {
                 var label = NP.Random(10);
                 inputs.Add(ToOneHot(label, oneHot));
             }
-            var outputs = new List<float>();
+            var outputs = new List<double>();
             inputs.ForEach(p =>
             {
                 outputs.AddRange(p);
@@ -137,9 +137,9 @@ namespace Engine.Brain.Entity
         /// </summary>
         /// <param name="inputs"></param>
         /// <returns></returns>
-        public static int Argmax(float[] inputs)
+        public static int Argmax(double[] inputs)
         {
-            float max = inputs.Max();
+            double max = inputs.Max();
             return Array.IndexOf(inputs, max);
         }
         /// <summary>
@@ -155,7 +155,7 @@ namespace Engine.Brain.Entity
             int[] output = new int[dim0];
             for (int i = 0; i < dim0; i++)
             {
-                float[] arr = new float[dim1];
+                double[] arr = new double[dim1];
                 for (int j = 0; j < dim1; j++)
                     arr[j] = inputs[i, j];
                 output[i] = Argmax(arr);
@@ -167,9 +167,9 @@ namespace Engine.Brain.Entity
         /// </summary>
         /// <param name="inputs"></param>
         /// <returns></returns>
-        public static float Max(float[] inputs)
+        public static double Max(double[] inputs)
         {
-            float max = inputs.Max();
+            double max = inputs.Max();
             return max;
         }
         /// <summary>
